Crate calypso_filety

calypso_filety

Binary file type definitions, parsers, high-level interfaces, and more for file types used in Calypso.

File types currently included are

Calypso Container File Format (CCFF). For more information, see the “spec”. Interfaces are located in the module ccff, with a high-level interface in ccff::hl and a lower-level (as the binary format with bincode) interface in ccff::ll.

Modules

ccff

Calypso Container File Format. See the module-level docs for more information.

ccff_new