Struct calypso_filety::ccff_new::ContainerFile

pub struct ContainerFile {
    abiver: u16,
    filety: u8,
    sections: IndexMap<Symbol, Section>,
}

Fields
abiver: u16filety: u8sections: IndexMap<Symbol, Section>
Implementations

impl ContainerFile

pub fn new(abiver: u16, filety: u8) -> Self

Create a new container file. The ABI version (abiver) and file type (filety) may be any arbitrary user-defined value.

pub fn set_abiver(&mut self, abiver: u16)

Set the ABI version of the container file. This may be any arbitrary user-defined value.

pub fn get_abiver(&self) -> u16

Get the ABI version of the container file.

pub fn set_filety(&mut self, filety: u8)

Set the file type of the container file. This may be any arbitrary user-defined value.

pub fn get_filety(&self) -> u8

Get the file type of the container file.

pub fn add_section(&mut self, name: Symbol, section: Section) -> Option<Section>

Add a section to the container file. If there was already a section with this name present, it will be replaced and returned.

Panics

This function will panic if the name of the section was longer than 255 characters, if the name of the section contained non-alphanumeric characters (including _), or if there were already 255 sections in the container file.

pub fn remove_section(&mut self, name: Symbol) -> Option<Section>

Remove a section from the container file. The removed section, if any, will be returned.

pub fn get_section(&self, name: Symbol) -> Option<&Section>

Get a reference to a section in the container file.

pub fn get_section_mut(&mut self, name: Symbol) -> Option<&mut Section>

Get a mutable reference to a section in the container file.

pub fn sections(&self) -> impl Iterator<Item = (&Symbol, &Section)>

Iterate over the sections in the container file.

pub fn sections_mut(&mut self) -> impl Iterator<Item = (&Symbol, &mut Section)>

Iterate mutably over the sections in the container file.

pub fn size(&self) -> usize

Get the size of the entire container file.

pub fn encode_to(self, buf: &mut Vec<u8>)

Encode this container file to the buffer provided. To allocate a sufficiently sized buffer, use Vec::with_capacity using the size given by ContainerFile::size.

Panics

This function will panic if a section was too large (larger than u32::MAX in bytes) or if there was too much data in the container file (due to architectural limitations, they are capped at around 4GiB)

pub fn encode(self) -> Vec<u8>